Отправка писем нескольким получателям переведена на параллельный режим, тесты успешно пройдены
